카테고리 그리드 기능 끝

+++ app/src/main/java/com/model/alchol_ranking/Review.kt
@@ -4,10 +4,11 @@ import com.google.gson.annotations.Expose
 
 import com.google.gson.annotations.SerializedName
 
+
 class Review {
     @SerializedName("score")
     @Expose
-    var score: Int? = null
+    var score: Float? = null
 
     @SerializedName("reviews")
     @Expose
